processImage/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
cachetools==5.3.0
decorator==5.1.1
frozenlist==1.3.3
fsspec==2023.1.0
gcloud==0.18.3
gcsfs==2023.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-documentai==2.10.0
google-cloud-storage==2.7.0
google-cloud-vision==3.3.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
install==1.3.5
multidict==6.0.4
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.3.5
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
google-cloud-tasks==2.12.1
